// rest-countries-api-with-color-theme-switcher-master/src/app/types/countryTypes.ts
export interface CountryDataType {
  name: {
    common: string;
    nativeName: {
      [key: string]: {
        common: string;
      };
    };
  };
  population: number;
  region: string;
  subregion: string;
  capital: string[];
  tld: string[];
  currencies: {
    [key: string]: {
      name: string;
    };
  };
  languages: {
    [key: string]: string;
  };
  borders: string[];
  cca3: string;
  flags: {
    png: string;
    alt: string;
  };
}